+++ docs/changelog.rst
@@ -30,6 +30,7 @@ Changelog
   used.  
 * :pull:`249` - Better sparse support for depletion matrix, ``depmtx`` files with a
   |depmtxReader|
+* :pull:`252` - Better axis and colorbar labeling for |detector| mesh plots
 
 
 Deprecations